+++ src/RejectedPromise.php
@@ -9,6 +9,8 @@ namespace GuzzleHttp\Promise;
  *
  * Thenning off of this promise will invoke the onRejected callback
  * immediately and ignore other callbacks.
+ *
+ * @final
  */
 class RejectedPromise implements PromiseInterface
 {