+++ src/TaskQueueInterface.php
@@ -8,19 +8,17 @@ interface TaskQueueInterface
 {
     /**
      * Returns true if the queue is empty.
-     *
-     * @return bool
      */
-    public function isEmpty();
+    public function isEmpty(): bool;
 
     /**
      * Adds a task to the queue that will be executed the next time run is
      * called.
      */
-    public function add(callable $task);
+    public function add(callable $task): void;
 
     /**
      * Execute all of the pending task in the queue.
      */
-    public function run();
+    public function run(): void;
 }